+++ practice/src/main/java/com/example/demo/admin/service/AdminService.java
@@ -1,11 +1,5 @@
 package com.example.demo.admin.service;
 
-import java.util.List;
-
-import com.example.demo.admin.data.ContactDataForAdmin;
-
 public interface AdminService {
-
-    List<ContactDataForAdmin> getAllContact();
 
 }
