+++ practice/src/main/java/com/example/demo/common/service/ContactService.java
@@ -1,8 +1,13 @@
 package com.example.demo.common.service;
 
+import java.util.List;
+
+import com.example.demo.admin.data.ContactDataForAdmin;
 import com.example.demo.contact.form.ContactForm;
 
 public interface ContactService {
 
     void saveContact(ContactForm contactForm);
+
+    List<ContactDataForAdmin> getAllContact();
 }
